Melhorando as notificações e incluindo confirmações de ação

+++ template/admin/pages/view.tpl.php
@@ -25,7 +25,7 @@
 
 <p>
     <a href="/admin/pages/1/edit" class="btn btn-primary">Editar</a>
-    <a href="/admin/pages/1/delete" class="btn btn-danger">Remover</a>
+    <a href="/admin/pages/1/delete" class="btn btn-danger confirm">Remover</a>
 </p>
 
 <hr>
